requirements.txt
streamlit>=1.30.0
google-genai>=0.16.0